виконані крок 1 та 2
