+++ routes/api/route.js
@@ -1,13 +1,8 @@
 // глобальні імпорти
 const express = require('express');
-const Joi = require('joi'); // метод валадації сталих виразів
-// локальні імпорти
-const {HttpError} = require('../../helpers'); // обробка помилок
 const controller = require('../../controler')
 
 const router = express.Router();
-
-
 
 
 router.get('/', controller.getAllContacts) 
@@ -21,6 +16,8 @@ router.delete('/:contactId', controller.deleteContact);
 router.put('/:contactId', controller.updatedContactById);
 
 router.patch('/:contactId/favorite', controller.favoritStatus);
+
+router.post('/:register', controller.userRegister)
